# shared/api-schemas/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Factory MES API",
    "version": "v1",
    "description": "Factory MES 시스템 REST API 계약 (OpenAPI 3.0.3)",
    "contact": {
      "name": "Factory MES Team",
      "email": "mes-team@example.com"
    }
  },
  "paths": {
...
    "/api/relationships": {
      "get": {
        "summary": "모든 relationship 목록 조회",
        "tags": [
          "Relationships"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RelationshipDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Relationship 생성",
        "tags": [
          "Relationships"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRelationshipRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelationshipDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/relationships/{id}": {
      "get": {
        "summary": "특정 relationship 조회",
        "tags": [
          "Relationships"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelationshipDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "summary": "특정 relationship 수정",
        "tags": [
          "Relationships"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRelationshipRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelationshipDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "summary": "특정 relationship 삭제",
        "tags": [
          "Relationships"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PropertyMapping": {
        "type": "object",
        "required": [
          "fromProperty",
          "toProperty"
        ],
        "properties": {
          "fromProperty": {
            "type": "string",
            "description": "source asset property key"
          },
          "toProperty": {
            "type": "string",
            "description": "target asset property key"
          },
          "transformRule": {
            "type": "string",
            "description": "Transform expression. Supported: 'value', 'value * N', 'value / N', 'value + N', 'value - N'",
            "default": "value"
          },
          "fromUnit": {
            "type": "string",
            "nullable": true
          },
          "toUnit": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "RelationshipDto": {
        "type": "object",
        "required": [
          "id",
          "fromAssetId",
          "toAssetId",
          "relationshipType",
          "properties",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "fromAssetId": {
            "type": "string"
          },
          "toAssetId": {
            "type": "string"
          },
          "relationshipType": {
            "type": "string"
          },
          "properties": {
            "type": "object",
            "additionalProperties": true
          },
          "mappings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyMapping"
            },
            "default": [],
            "description": "Property mappings for Supplies propagation. When empty, legacy properties.transfers TransferSpec is used."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CreateRelationshipRequest": {
        "type": "object",
        "required": [
          "fromAssetId",
          "toAssetId",
          "relationshipType"
        ],
        "properties": {
          "fromAssetId": {
            "type": "string"
          },
          "toAssetId": {
            "type": "string"
          },
          "relationshipType": {
            "type": "string"
          },
          "properties": {
            "type": "object",
            "additionalProperties": true
          },
          "mappings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyMapping"
            },
            "default": [],
            "description": "Optional property mappings; empty uses transfers-only behavior."
          }
        }
      },
      "UpdateRelationshipRequest": {
        "type": "object",
        "properties": {
          "fromAssetId": {
            "type": "string",
            "nullable": true
          },
          "toAssetId": {
            "type": "string",
            "nullable": true
          },
          "relationshipType": {
            "type": "string",
            "nullable": true
          },
          "properties": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true
          },
          "mappings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyMapping"
            },
            "nullable": true,
            "description": "When omitted, existing mappings are unchanged."
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}